requirements.txt
Cython==0.25.2
easydict==1.7
numpy==1.12.1
opencv-python==3.2.0.7
pandas==0.19.2
Pillow==4.1.1
PyYAML==3.12
scipy==0.19.0
tensorflow-gpu==1.1.0
tqdm==4.11.2